moved model_db to worker
