+++ pyproject.toml
@@ -74,6 +74,3 @@ mkdocs-glightbox = "^0.3.7"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.scripts]
-download-model = 'download_embedder:download'
