+++ worker/Dockerfile
@@ -9,7 +9,6 @@ ENV POETRY_NO_INTERACTION=1 \
 
 WORKDIR /app/
 ADD pyproject.toml poetry.lock ./
-ADD download_embedder.py /app/
 
 RUN --mount=type=cache,target=$POETRY_CACHE_DIR poetry install --no-root --no-ansi --with worker,api --without ai,dev
 
@@ -29,11 +28,9 @@ COPY --from=builder ${VIRTUAL_ENV} ${VIRTUAL_ENV}
 WORKDIR /app/
 ADD redbox/ /app/redbox
 ADD worker/ /app/worker
-ADD download_embedder.py /app/
-ADD worker/download_ocr_models.py /app/
 
 # Download the model
-RUN python download_embedder.py --embedding_model ${EMBEDDING_MODEL}
+RUN python worker/download_embedder.py --embedding_model ${EMBEDDING_MODEL}
 
 # Downlaod Sentence Tokenizer
 RUN python -m nltk.downloader punkt
@@ -42,6 +39,6 @@ RUN python -m nltk.downloader punkt
 RUN python -m nltk.downloader averaged_perceptron_tagger
 
 # Download the OCR models
-RUN python download_ocr_models.py
+RUN python worker/download_ocr_models.py
 
 CMD ["uvicorn", "worker.src.app:app", "--host", "0.0.0.0", "--port", "5000"]
