--- worker/download_ocr_models.py
@@ -1,4 +1,0 @@
-from unstructured_inference.models.tables import load_agent
-
-if __name__ == "__main__":
-    load_agent()
